Support deserializing sequences from non-contiguous elements
The sequence deserializer had been modified to read all XML events until the end of the current depth. Unused events are kept in a buffer which is re-read by the parent deserializer.

The buffer will be trimmed when the top-level deserializer moves on to avoid excess memory usage. Events which have been consumed by sequence deserializers in the middle of the buffer are replaced with tombstones to avoid unnecessary shuffling.

Some documentation which isn't strictly relevant to this change has been added in a few modules which I found hard to follow at first.

A few considerations:

- `Deserializer` type signature has an unnecessary type param `R: Read`, kept to avoid an API change. But this could be removed.
- Reading of out-of-order elements currently cannot be configured or disabled. This has some consequences:
    - Some callers may have previously relied on errors produced by non-contiguous elements, particularly for tuples.
    - This may be a memory usage concern for large documents.